Render Background

+++ AA2_Arkanoid/Vec2.cpp
@@ -5,7 +5,7 @@
 Vec2::Vec2() : x(0), y(0) {}
 
 //Parameters constructor
-Vec2::Vec2(const float _x, const float _y) : x(_x), y(_y) {}
+Vec2::Vec2(const int _x, const int _y) : x(_x), y(_y) {}
 
 //Copy constructor
 Vec2::Vec2(const Vec2 &other) : x(other.x), y(other.y) {}
