+++ AA2_Arkanoid/Vec2.h
@@ -8,7 +8,7 @@ public:
 
 //Constructors
 	Vec2();
-	Vec2(const float, const float);
+	Vec2(const int, const int);
 	Vec2(const Vec2 &);
 
 //Destructor
